+++ prisma/migrations/20250906115030_added_symbol_profile_relation_to_market_data/migration.sql
@@ -1,0 +1,8 @@
+-- AlterTable
+ALTER TABLE "public"."MarketData" ADD COLUMN "symbolProfileId" TEXT;
+
+-- CreateIndex
+CREATE INDEX "MarketData_symbolProfileId_idx" ON "public"."MarketData"("symbolProfileId");
+
+-- AddForeignKey
+ALTER TABLE "public"."MarketData" ADD CONSTRAINT "MarketData_symbolProfileId_fkey" FOREIGN KEY ("symbolProfileId") REFERENCES "public"."SymbolProfile"("id") ON DELETE CASCADE ON UPDATE CASCADE;